apiserver/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
black==24.8.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-yasg==1.21.7
eventlet==0.37.0
frozenlist==1.4.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
inflection==0.5.1
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langsmith==0.1.128
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.49.0
orjson==3.10.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
psycopg2==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
pypdf==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.31.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
yarl==1.12.1
